python code refactoring
